Add some improvements

+++ src/main/java/fr/polytech/cloud/entities/PositionMongoDBEntity.java
@@ -2,11 +2,12 @@ package fr.polytech.cloud.entities;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public @Data class PositionMongoDBEntity {
 
     private String type;
 
-    private List<Double> coordinates;
+    private List<BigDecimal> coordinates;
 }
